update new command
