+++ README.md
@@ -44,7 +44,7 @@ The `new` command creates a new Koop project from the template at the current lo
 koop new <type> <name>
 
 Positionals:
-  type  project type                       [string] [choices: "app", "provider"]
+  type  project type               [string] [choices: "app", "provider", "auth"]
   name  project name                                                    [string]
 
 Provider Options:
